requirements.txt
requests
transformers
torch
Pillow
python-dotenv
streamlit